move trie checker into its own file
